Extract Histogram's settings bar into a shared PlotSettingsBar component
Moves the download button, "add region" search popup, and transpose-aware
container styling out of Histogram-specific HistogramSettings into a
reusable PlotSettingsBar in plots-general.tsx, alongside the other shared
plot infrastructure. Histogram-specific controls (type dropdown, relative
checkbox) are now passed as children. Behavior-preserving -- same DOM
structure and rendering order as before.
